we might be cooked
